+++ modules/item/proficiency-item.js
@@ -1,4 +1,5 @@
-import { HMItem } from './item.js';
+import { HMItem } from "./item.js";
+import { HMAggregator } from "../rules/aggregator.js";
 
 export class HMProficiencyItem extends HMItem {
     prepareBaseData() {
@@ -8,27 +9,28 @@ export class HMProficiencyItem extends HMItem {
 
     prepareDerivedData() {
         super.prepareDerivedData();
+        // We're reading vectors from a getter, so pathing will be invalid.
+        this.hmagg = new HMAggregator({ parent: this }, { noprop: false, readonly: true });
+    }
+
+    /**
+     * @param {HMAggregator} aggregator
+     */
+    _postAggregation(aggregator) {
+        const { system } = this;
+        const [isMechanical, isRanged] = [system.mechanical.checked, system.ranged.checked];
+
+        const speedUnits = aggregator.getUnitsForStat("spd");
+        speedUnits.forEach(unit => unit > 0 && (unit.value = -unit));
+
+        if (isRanged) aggregator.deleteUnitsByStat("def");
+        if (isMechanical) aggregator.deleteUnitsByStat("dmg");
     }
 
     _prepProficiencyData() {
-        const {system} = this;
-        const {bonus} = system;
-        const [isMechanical, isRanged] = [system.mechanical.checked, system.ranged.checked];
-
-        let dirty = false;
-        if (bonus.spd > 0) {
-            bonus.spd = -Math.abs(bonus.spd);
-            dirty = true;
+        const { system } = this;
+        if (system.mechanical.checked && !system.ranged.checked) {
+            this.update({ "system.mechanical.checked": false });
         }
-
-        if (isMechanical || isRanged) {
-            if (isMechanical && !isRanged) this.update({'system.mechanical.checked': false});
-
-            const {def, dmg} = bonus;
-            if (isRanged) bonus.def = 0;
-            if (isMechanical) bonus.dmg = 0;
-            dirty = dirty || bonus.def !== def || bonus.dmg !== dmg;
-        }
-        if (dirty) this.update({'system.bonus': bonus});
     }
 }